Implement protocol version 1. This is not compatible with version 0.  Uses 24-byte messages and doesn't distinguish routes from xroutes.

+++ kernel.c
@@ -96,7 +96,7 @@ write_proc(char *filename, int value)
 }
 
 struct netlink {
-    unsigned int seqno;
+    unsigned short seqno;
     int sock;
     struct sockaddr_nl sockaddr;
     socklen_t socklen;
